Code refactoring and security
The functionality of the website is unchanged, but this should increase security and the code should be more standardized and readable (still some changes I would like to make but this is an improvement).

config.php: always "require_once config.php" instead of using session_start(), this creates a more secure session.

logout.php: instead of handling logout with forms, simply redirect to this page and it will empty all session variables and destroy the session.

SQL: sql files are stored here just so the file system isn't so cluttered